Master - CXXX - CPF com zero na frente II
	ACTWEB 29.714.789.61

+++ branches/LFSistemas.VLI.ACTWeb.DataAccessObjects/IDDAO.cs
@@ -43,7 +43,7 @@ namespace LFSistemas.VLI.ACTWeb.DataAccessObjects
                     {
                         while (reader.Read())
                         {
-                            retorno = reader.GetDouble(0);
+                            retorno = (double)reader.GetDecimal(0) ;
                         }
                     }
